Add docker diskio stats on Windows
