+++ metricbeat/module/docker/diskio/helper_unix.go
@@ -1,0 +1,120 @@
+// Licensed to Elasticsearch B.V. under one or more contributor
+// license agreements. See the NOTICE file distributed with
+// this work for additional information regarding copyright
+// ownership. Elasticsearch B.V. licenses this file to you under
+// the Apache License, Version 2.0 (the "License"); you may
+// not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
+// +build !windows
+
+package diskio
+
+import (
+	"time"
+
+	"github.com/docker/docker/api/types"
+
+	"github.com/elastic/beats/metricbeat/module/docker"
+)
+
+// BlkioService is a helper to collect and calculate disk I/O metrics
+type BlkioService struct {
+	lastStatsPerContainer map[string]BlkioRaw
+}
+
+// NewBlkioService builds a new initialized BlkioService
+func NewBlkioService() *BlkioService {
+	return &BlkioService{
+		lastStatsPerContainer: make(map[string]BlkioRaw),
+	}
+}
+
+func (io *BlkioService) getBlkioStatsList(rawStats []docker.Stat, dedot bool) []BlkioStats {
+	formattedStats := []BlkioStats{}
+
+	statsPerContainer := make(map[string]BlkioRaw)
+	for _, myRawStats := range rawStats {
+		stats := io.getBlkioStats(&myRawStats, dedot)
+		statsPerContainer[myRawStats.Container.ID] = stats.serviced
+		formattedStats = append(formattedStats, stats)
+	}
+
+	io.lastStatsPerContainer = statsPerContainer
+	return formattedStats
+}
+
+func (io *BlkioService) getBlkioStats(myRawStat *docker.Stat, dedot bool) BlkioStats {
+	newBlkioStats := io.getNewStats(myRawStat.Stats.Read, myRawStat.Stats.BlkioStats.IoServicedRecursive)
+	bytesBlkioStats := io.getNewStats(myRawStat.Stats.Read, myRawStat.Stats.BlkioStats.IoServiceBytesRecursive)
+
+	myBlkioStats := BlkioStats{
+		Time:      myRawStat.Stats.Read,
+		Container: docker.NewContainer(myRawStat.Container, dedot),
+
+		serviced:      newBlkioStats,
+		servicedBytes: bytesBlkioStats,
+	}
+
+	oldBlkioStats, exist := io.lastStatsPerContainer[myRawStat.Container.ID]
+	if exist {
+		myBlkioStats.reads = io.getReadPs(&oldBlkioStats, &newBlkioStats)
+		myBlkioStats.writes = io.getWritePs(&oldBlkioStats, &newBlkioStats)
+		myBlkioStats.totals = io.getTotalPs(&oldBlkioStats, &newBlkioStats)
+	}
+
+	return myBlkioStats
+}
+
+func (io *BlkioService) getNewStats(time time.Time, blkioEntry []types.BlkioStatEntry) BlkioRaw {
+	stats := BlkioRaw{
+		Time:   time,
+		reads:  0,
+		writes: 0,
+		totals: 0,
+	}
+
+	for _, myEntry := range blkioEntry {
+		switch myEntry.Op {
+		case "Write":
+			stats.writes += myEntry.Value
+		case "Read":
+			stats.reads += myEntry.Value
+		case "Total":
+			stats.totals += myEntry.Value
+		}
+	}
+	return stats
+}
+
+func (io *BlkioService) getReadPs(old *BlkioRaw, new *BlkioRaw) float64 {
+	duration := new.Time.Sub(old.Time)
+	return calculatePerSecond(duration, old.reads, new.reads)
+}
+
+func (io *BlkioService) getWritePs(old *BlkioRaw, new *BlkioRaw) float64 {
+	duration := new.Time.Sub(old.Time)
+	return calculatePerSecond(duration, old.writes, new.writes)
+}
+
+func (io *BlkioService) getTotalPs(old *BlkioRaw, new *BlkioRaw) float64 {
+	duration := new.Time.Sub(old.Time)
+	return calculatePerSecond(duration, old.totals, new.totals)
+}
+
+func calculatePerSecond(duration time.Duration, old uint64, new uint64) float64 {
+	value := float64(new) - float64(old)
+	if value < 0 {
+		value = 0
+	}
+	return value / duration.Seconds()
+}
